% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conProb.R
\name{conProb}
\alias{conProb}
\title{Consistency probability calculation for one MRCT}
\usage{
conProb(alpha, power, pi, regionFraction)
}
\arguments{
\item{alpha}{The type I error.}

\item{power}{Power.}

\item{pi}{The threshold ratio in Japan's criterion I.}

\item{regionFraction}{The regional fraction.}
}
\value{
consistencyProbability The consistency probability.
}
\description{
conProb returns the consistency probability given the fraction calculation
}
\examples{
conProb(0.025,0.8,0.5,0.23)
}
