% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conProb2MRCTs.R
\name{conProb2MRCTs}
\alias{conProb2MRCTs}
\title{Consistency probability calculation for two MRCTs}
\usage{
conProb2MRCTs(
  alpha,
  power,
  pi,
  regionFraction1,
  regionFraction2,
  d1,
  d2,
  sigmaTrt1,
  sigmaCtrl1,
  sigmaTrt2,
  sigmaCtrl2,
  ratio
)
}
\arguments{
\item{alpha}{The type I error.}

\item{power}{Power.}

\item{pi}{The threshold ratio in Japan's criterion I.}

\item{regionFraction1}{The regional fraction for MRCT 1.}

\item{regionFraction2}{The regional fraction for MRCT 2.}

\item{d1}{The true mean of difference of response for MRCT 1}

\item{d2}{The true mean of difference of response for MRCT 2}

\item{sigmaTrt1}{The standard deviation of response in the treatment group for MRCT1}

\item{sigmaCtrl1}{The standard deviation of response in the control group for MRCT1}

\item{sigmaTrt2}{The standard deviation of response in the treatment group for MRCT2}

\item{sigmaCtrl2}{The standard deviation of response in the control group for MRCT2}

\item{ratio}{The ratio between the treatment group and control group}
}
\value{
consistencyProbability The consistency probability.
}
\description{
Consistency probability calculation for two MRCTs
}
\examples{
conProb2MRCTs(0.025,0.8,0.5,0.127,0.127,1,1,4,4,4,4,1)
}
