% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conProb2MRCTsH.R
\name{conProb2MRCTsH}
\alias{conProb2MRCTsH}
\title{Regional fraction calculation for two homogeneous MRCTs when the regional fractions are equal}
\usage{
conProb2MRCTsH(alpha, power, pi, regionFraction1, regionFraction2)
}
\arguments{
\item{alpha}{The type I error.}

\item{power}{Power.}

\item{pi}{The threshold ratio in Japan's criterion I.}

\item{regionFraction1}{The regional fraction for MRCT 1.}

\item{regionFraction2}{The regional fraction for MRCT 2.}
}
\value{
consistencyProbability The consistency probability.
}
\description{
Regional fraction calculation for two homogeneous MRCTs when the regional fractions are equal
}
\examples{
conProb2MRCTsH(0.025,0.8,0.5,0.127,0.127)
}
