% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reFrac.R
\name{reFrac}
\alias{reFrac}
\title{Regional fraction calculation for one MRCT}
\usage{
reFrac(alpha, power, pi, consistencyProbability)
}
\arguments{
\item{alpha}{The type I error.}

\item{power}{Power.}

\item{pi}{The threshold ratio in Japan's criterion I.}

\item{consistencyProbability}{The consistency probability.}
}
\value{
regionFraction The regional fraction.
}
\description{
reFrac returns the fraction calculation given the consistency probability
}
\examples{
reFrac(0.025,0.8,0.5,0.8)
}
