% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reFrac2MRCTs.R
\name{reFrac2MRCTs}
\alias{reFrac2MRCTs}
\title{Regional fraction calculation for two homogeneous MRCTs when the regional fractions are equal}
\usage{
reFrac2MRCTs(alpha, power, pi, consistencyProbability)
}
\arguments{
\item{alpha}{The type I error.}

\item{power}{Power.}

\item{pi}{The threshold ratio in Japan's criterion I.}

\item{consistencyProbability}{The consistency probability.}
}
\value{
regionFraction The regional fraction.
}
\description{
Regional fraction calculation for two homogeneous MRCTs when the regional fractions are equal
}
\examples{
reFrac2MRCTs(0.025,0.8,0.5,0.8)
}
