% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table2MRCTs}
\alias{table2MRCTs}
\title{Regional fraction calculation for one MRCT}
\usage{
table2MRCTs(numSimu, choice, seed)
}
\arguments{
\item{numSimu}{The number of simulation times.}

\item{choice}{Table Number.}

\item{seed}{Random seed}
}
\value{
Tables
}
\description{
table2MRCTs returns the tables shown in the reference
}
\examples{
table2MRCTs(100000,1,1)
}
